# Put coordinate in tuple instead of dict