Hi this PR improves the performance by replacing dict with tuple to store coordinates.

The profiling shows a 2X improvement on encoding linestrings.

Before the PR:
```
         4551974 function calls (4549470 primitive calls) in 15.246 seconds

   Ordered by: standard name

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
...

      500    0.011    0.000   11.351    0.023 encoder.py:205(addFeature)
      500    0.005    0.000    8.267    0.017 encoder.py:305(_parseGeometry)
    50000    0.150    0.000    0.209    0.000 encoder.py:310(_get_point_obj)
      500    0.632    0.001    8.246    0.016 encoder.py:318(_get_arc_obj)
...
```

After this PR:
```
         2660437 function calls (2657933 primitive calls) in 7.743 seconds

   Ordered by: standard name

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
...
      500    0.009    0.000    3.945    0.008 encoder.py:205(addFeature)
      500    0.005    0.000    0.909    0.002 encoder.py:305(_parseGeometry)
      500    0.154    0.000    0.889    0.002 encoder.py:310(_get_arc_obj)
      500    1.088    0.002    3.888    0.008 encoder.py:356(_geo_encode)
...
```

The benchmark script is here: https://gist.github.com/ptpt/1bf2351e91ef6b9b3eac23829ccbdb2b